Added missing canvas in the case of video-output
For the video to be save-able, you need to define the canvas
before the FunctionAnimation, but you may only do so for the video
output, not for the regular plotting-window output. (Why that is,
is beyond me but does not seem worth investigating right now.)

Also, while at it, cleaned up the code a bit and removed unused imports.